Refactor out `analy` and `analy_alt`
Refactors classes on `analy` and `analy_alt` out to `movements` and
`guis` modules. Consolidates the Trend classes that remain to
`analy` module, as `Trends` and `TrendAlt`. Removes `analy_alt`.
