+++ src/market_analy/trends/__init__.py
@@ -12,10 +12,12 @@ import pandas as pd
 
 # necessary for unpickling test resources. When picked analy was here
 # ... i.e. at market_analy.trend. Also serves to provide legacy access.
-from .analy import Movement, Movements, Trends, TrendsGui  # noqa: F401
+from .analy import Trends  # noqa: F401
+from .guis import TrendsGui  # noqa: F401
+from .movements import Movement, Movements  # noqa: F401
 
 if typing.TYPE_CHECKING:
-    from .movements_base import MovementsSupportChartAnaly
+    from .movements import MovementsSupportChartAnaly
 
 
 class TrendsProto(typing.Protocol):
@@ -24,5 +26,15 @@ class TrendsProto(typing.Protocol):
     data: pd.DataFrame
     interval: mp.intervals.RowInterval
 
+    def __init__(
+        self,
+        data: pd.DataFrame,
+        interval: mp.intervals.RowInterval,
+        *args: typing.Any,
+        **kwargs: typing.Any,
+    ):
+        ...
+
     def get_movements(self) -> MovementsSupportChartAnaly:
         """Evaluate all movements over `data`."""
+        ...
