+++ src/market_analy/trends/guis.py
@@ -1,26 +1,29 @@
-"""Guis for trend analysis."""
+"""GUIs to visualise movements identifeid by Trends classes."""
 
 from __future__ import annotations
 
+import typing
+from collections import abc
 from collections.abc import Callable
 from contextlib import contextmanager
-from typing import TYPE_CHECKING
 
 import bqplot as bq
 import ipyvuetify as v
 import market_prices as mp
 import pandas as pd
 
+import market_analy
+
 from ..guis import ChartOHLCCaseBase
 from ..utils import ipyvuetify_utils as vu
 from ..utils.bq_utils import HFixedRule
 from . import charts
+from .analy import RvrAlt, Trends, TrendsAlt
 
-if TYPE_CHECKING:
+if typing.TYPE_CHECKING:
     from .. import analysis as ma_analysis
     from . import TrendsProto
-    from .analy import Movement as Movement
-    from .analy_alt import Movement as MovementAlt
+    from .movements import Movement, MovementAlt, Movements, MovementsAlt
 
 
 class TrendsGuiBase(ChartOHLCCaseBase):
@@ -147,7 +150,7 @@ class TrendsGuiBase(ChartOHLCCaseBase):
     def current_case(self) -> Movement | MovementAlt | None:
         """Current selected case"""
         case = super().current_case
-        if TYPE_CHECKING:
+        if typing.TYPE_CHECKING:
             assert isinstance(case, (Movement, MovementAlt))
         return case
 
@@ -221,3 +224,173 @@ class TrendsGuiBase(ChartOHLCCaseBase):
         self.chart.reset_x_ticks()
         self._set_slider_limits_to_all_plottable_x_ticks()
         self.chart.update_trend_mark()
+
+
+class TrendsGui(TrendsGuiBase):
+    """GUI to visualise movements evaluated by `analy.Trends` class.
+
+    Parameters
+    ----------
+    `prd`, `ext_break`, `ext_limit` and `min_bars` all as `analy.Trends`
+    class.
+
+    All other parameters as base class `TrendsGuiBase`.
+    """
+
+    def __init__(
+        self,
+        analysis: ma_analysis.Analysis,
+        interval: mp.intervals.RowInterval,
+        prd: int,
+        ext_break: float,
+        ext_limit: float,
+        min_bars: int = 0,
+        max_ticks: int | None = None,
+        log_scale: bool = True,
+        display: bool = True,
+        **kwargs,
+    ):
+        trend_kwargs = {
+            "prd": prd,
+            "ext_break": ext_break,
+            "ext_limit": ext_limit,
+            "min_bars": min_bars,
+        }
+        super().__init__(
+            analysis=analysis,
+            interval=interval,
+            trend_cls=Trends,
+            trend_kwargs=trend_kwargs,
+            max_ticks=max_ticks,
+            log_scale=log_scale,
+            display=display,
+            narrow_view=prd,
+            wide_view=prd * 3,
+            **kwargs,
+        )
+        self.cases: Movements
+        self.trends: Trends
+
+    def _gui_click_case_handler(self, mark: bq.Scatter, event: dict):
+        """Gui level handler for clicking a mark representing a trend start.
+
+        Lightens 'show all scatters' button to indicate option available.
+        Displays tooltip to html output.
+        """
+        self.cases_controls_container.lighten_single_case()
+        self.cases_controls_container.but_show_all.darken()
+        move = self.cases.event_to_case(mark, event)
+        html = self.cases.get_move_html(move)
+        self.html_output.display(html)
+
+    @property
+    def current_case(self) -> Movement | None:
+        """Current selected case"""
+        # method included to update type to Movement class as defined on this module
+        case = super().current_case
+        if typing.TYPE_CHECKING:
+            assert isinstance(case, Movement)
+        return case
+
+    def _add_rulers(self):
+        self._close_rulers()  # close any existing
+        case = self.current_case
+        assert case is not None
+        is_adv = case.is_adv
+
+        ohlc_mark = next(m for m in self.chart.figure.marks if isinstance(m, bq.OHLC))
+
+        fctrs = self.trends.fctrs_pos_break if is_adv else self.trends.fctrs_neg_break
+        self._rulers.append(
+            market_analy.utils.bq_utils.TrendRule(
+                x=case.start.asm8,
+                y=case.start_px,
+                length=case.params["prd"],
+                factors=fctrs,
+                scales=self.chart.scales,
+                ordinal_values=list(ohlc_mark.x),
+                figure=self.chart.figure,
+                color="orange",
+                draggable=True,
+                stroke_width=3,
+            )
+        )
+
+        fctrs = self.trends.fctrs_pos_limit if is_adv else self.trends.fctrs_neg_limit
+        idx = -case.params["prd"]
+        self._rulers.append(
+            market_analy.utils.bq_utils.TrendRule(
+                x=case.line_limit.index[idx].asm8,
+                y=case.line_limit.iloc[idx],
+                length=case.params["prd"],
+                factors=fctrs,
+                scales=self.chart.scales,
+                ordinal_values=list(ohlc_mark.x),
+                figure=self.chart.figure,
+                color="skyblue",
+                draggable=True,
+                stroke_width=3,
+            )
+        )
+
+
+class TrendsAltGui(TrendsGuiBase):
+    """GUI to visualise movements evaluated by `analy_alt.Trends` class.
+
+    Parameters
+    ----------
+    `prd`, `ext`, `rvr`, `grad`, `rvr_init` and `min_bars` all as
+    `analy_alt.Trends` class.
+
+    All other parameters as base class `TrendsGuiBase`.
+    """
+
+    def __init__(
+        self,
+        analysis: ma_analysis.Analysis,
+        interval: mp.intervals.RowInterval,
+        prd: int,
+        ext: float,
+        rvr: float | abc.Sequence[float] | RvrAlt,
+        grad: float | None,
+        rvr_init: float | tuple[float, float] | list[float],
+        min_bars: int,
+        max_ticks: int | None = None,
+        log_scale: bool = True,
+        display: bool = True,
+        **kwargs,
+    ):
+        trend_kwargs = {
+            "prd": prd,
+            "ext": ext,
+            "grad": grad,
+            "rvr": rvr,
+            "rvr_init": rvr_init,
+            "min_bars": min_bars,
+        }
+        super().__init__(
+            analysis=analysis,
+            interval=interval,
+            trend_cls=TrendsAlt,
+            trend_kwargs=trend_kwargs,
+            max_ticks=max_ticks,
+            log_scale=log_scale,
+            display=display,
+            narrow_view=prd,
+            wide_view=prd * 3,
+            **kwargs,
+        )
+        self.cases: MovementsAlt
+        self.trends: TrendsAlt
+
+    def _gui_click_case_handler(self, mark: bq.Scatter, event: dict):
+        """Gui level handler for clicking a mark representing a trend start.
+
+        Lightens 'show all scatters' button to indicate option available.
+        Displays tooltip to html output.
+        """
+        self.cases_controls_container.lighten_single_case()
+        self.cases_controls_container.but_show_all.darken()
+        move = self.cases.event_to_case(mark, event)
+        html = self.cases.get_move_html(move)
+        self.html_output.display(html)
